+++ LabXml/Validator/Tfs/TfsSqlIsosExist.cs
@@ -58,7 +58,8 @@ namespace AutomatedLab
                         break;
                     case Roles.AzDevOps:
                         requiredRoles.Add("2017");
-                        sqlmachines.AddRange(lab.Machines.Where(m => m.Roles.Where(r => r.Name == Roles.SQLServer2017).Count() > 0).Select(m => m.Name));
+                        requiredRoles.Add("2019");
+                        sqlmachines.AddRange(lab.Machines.Where(m => m.Roles.Where(r => r.Name == Roles.SQLServer2017 || r.Name == Roles.SQLServer2019).Count() > 0).Select(m => m.Name));
                         break;
                     default:
                         break;